refactor: migrate settings to their own pages
